feat: add comprehensive tests for Stmt2 decimal binding behavior
